moved onFormSubmit function up to container and got it working again. next up, connecting it to dropDown component
